glib: Fix using libffi exports.
The default upstream glib project expects it to be linked with libffi as a static library. So it defines FFI_BUILDING which in turn defines the imports (like ffi_type_void) in ffi.h as `extern`.

Since we now build libffi as a dynamic library, gobject ends up being compiled as if it has its own copies of all those exports, but in reality they reference garbage addresses and blow up when used.

We need to remove this flag. Then the imports are compiled as `extern __declspec(dllimport)` and correctly linked to libffi.dll's exports at runtime.